refactor: . move new  postpone tests from  QueryRendererHelper.test.ts to Postponer.test.ts
